# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Integrador API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alertas"
                ],
                "summary": "Lista todas las alertas",
                "description": "Obtiene una lista de todas las alertas disponibles.",
                "operationId": "a4d0d36b07da650ba1d6b0d78f06a7f5",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Alert"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Alertas"
                ],
                "summary": "Crea una nueva alerta",
                "description": "Crea una nueva alerta con la información proporcionada.",
                "operationId": "eb1fb84f3bf66ec5e88133c5b8e05c31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alert"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alertas"
                ],
                "summary": "Obtiene una alerta por ID",
                "description": "Obtiene los detalles de una alerta específica por su ID.",
                "operationId": "3e280f1020a68722f5bcb7a394532cc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alert"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Alertas"
                ],
                "summary": "Actualiza una alerta por ID",
                "description": "Actualiza la información de una alerta específica por su ID.",
                "operationId": "d2bdb2e600ed301dde96d5f3c82e8c81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alert"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alertas"
                ],
                "summary": "Elimina una alerta por ID",
                "description": "Elimina una alerta específica por su ID.",
                "operationId": "93ab7eb38374b1b38e4e24ce70f08a8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operación exitosa (sin contenido)"
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreAlertRequest": {
                "title": "Store Alert Request",
                "description": "Store Alert request body data",
                "required": [
                    "recurrencia",
                    "tipo",
                    "subtipo",
                    "operador_id",
                    "paciente_id"
                ],
                "properties": {
                    "recurrencia": {
                        "description": "Alert recurrence (puntual/periodic)",
                        "type": "string",
                        "enum": [
                            "puntual",
                            "periodic"
                        ]
                    },
                    "tipo": {
                        "description": "Alert type",
                        "type": "string",
                        "enum": [
                            "avisos",
                            "protocolos",
                            "ausencias_retornos"
                        ]
                    },
                    "subtipo": {
                        "description": "Alert subtype",
                        "type": "string"
                    },
                    "fecha": {
                        "description": "Alert date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "dia_semana": {
                        "description": "Day of the week (1-7)",
                        "type": "integer",
                        "nullable": true
                    },
                    "descripcion": {
                        "description": "Alert description",
                        "type": "string",
                        "nullable": true
                    },
                    "operador_id": {
                        "description": "Operator ID",
                        "type": "integer"
                    },
                    "paciente_id": {
                        "description": "Patient ID",
                        "type": "integer"
                    },
                    "zona_id": {
                        "description": "Zone ID",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateAlertRequest": {
                "title": "Update Alert Request",
                "description": "Update Alert request body data",
                "properties": {
                    "recurrencia": {
                        "description": "Alert recurrence (puntual/periodic)",
                        "type": "string",
                        "enum": [
                            "puntual",
                            "periodic"
                        ]
                    },
                    "tipo": {
                        "description": "Alert type",
                        "type": "string",
                        "enum": [
                            "avisos",
                            "protocolos",
                            "ausencias_retornos"
                        ]
                    },
                    "subtipo": {
                        "description": "Alert subtype",
                        "type": "string"
                    },
                    "fecha": {
                        "description": "Alert date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "dia_semana": {
                        "description": "Day of the week (1-7)",
                        "type": "integer",
                        "nullable": true
                    },
                    "descripcion": {
                        "description": "Alert description",
                        "type": "string",
                        "nullable": true
                    },
                    "operador_id": {
                        "description": "Operator ID",
                        "type": "integer"
                    },
                    "paciente_id": {
                        "description": "Patient ID",
                        "type": "integer"
                    },
                    "zona_id": {
                        "description": "Zone ID",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Alert": {
                "title": "Alert",
                "description": "Alert model",
                "properties": {
                    "id": {
                        "description": "Alert ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "message": {
                        "description": "Alert message",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}